requirements.txt
Django~=5.0.6
selenium==4.20.0
djangorestframework==3.15.1
